requirements.txt
pdfkit
pyyaml
networkx
matplotlib
# tabulate
